Provide (Java) API to create a DataFrame from an RDD[GenericRecord]
A higher level API provides more flexibility for implementation.
The API can now be invoked from Java as well, by means of the new AvroUtil object.
